Avoid dep on from_native.dart
