+++ packages/realm_dart/lib/src/configuration.dart
@@ -8,7 +8,6 @@ import 'dart:io';
 import 'package:path/path.dart' as _path;
 
 import 'app.dart';
-import 'handles/from_native.dart';
 import 'handles/realm_core.dart';
 import 'init.dart';
 import 'logging.dart';
@@ -712,6 +711,31 @@ final class CompensatingWriteError extends SyncError {
   String toString() {
     return "CompensatingWriteError: $message. ${compensatingWrites ?? ''}";
   }
+}
+
+/// @nodoc
+class SyncErrorDetails {
+  final String message;
+  final SyncErrorCode code;
+  final String? path;
+  final bool isFatal;
+  final bool isClientResetRequested;
+  final String? originalFilePath;
+  final String? backupFilePath;
+  final List<CompensatingWriteInfo>? compensatingWrites;
+  final Object? userError;
+
+  SyncErrorDetails(
+    this.message,
+    this.code,
+    this.userError, {
+    this.path,
+    this.isFatal = false,
+    this.isClientResetRequested = false,
+    this.originalFilePath,
+    this.backupFilePath,
+    this.compensatingWrites,
+  });
 }
 
 /// @nodoc
